Makefile edits for FreeBSD usage
These edits do not use anything to make them portable,
they have been just written to support the platform I'm building on
(FreeBSD10 amd64). Please keep in mind that you will have to modify
the config.mk and possibly the Makefile as well to get it to build
on your platform.